requirements.txt
certifi==2018.11.29
chardet==3.0.4
construct==2.5.3
Django==1.11.11
django-ckeditor==5.6.1
django-crontab==0.7.1
django-haystack==2.8.1
django-ipware==2.1.0
django-js-asset==1.1.0
django-qiniu-storage==2.3.1
djangorestframework==3.9.1
drf-haystack==1.8.4
elasticsearch==6.3.1
feedparser==5.2.1
future==0.17.1
httplib2==0.12.0
idna==2.8
Markdown==3.0.1
mistune==0.8.4
oauth2client==4.1.3
pefile==2018.8.8
Pillow==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
Pygments==2.3.1
PyMySQL==0.9.3
python-dateutil==2.7.5
python-ptrace==0.9.3
pytz==2018.9
qiniu==7.2.2
redis==3.0.1
requests==2.21.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.17.1
